working with categories table, mixins, schemas and linters
