+++ docker-compose.yml
@@ -23,10 +23,6 @@ services:
     volumes:
       - postgres-data:/var/lib/postgresql/data/
       - ./database:/var/lib/postgresql/data/
-      # - ./data_exports:/var/lib/postgresql/exports/
-      # - ./init-development-db.sh:/docker-entrypoint-initdb.d/init-development-db.sh
-      # - ./permissions.sh:/docker-entrypoint-initdb.d/permissions.sh
-      # - ./init.sql:/docker-entrypoint-initdb.d/init.sql
       - shared-backups:/var/lib/postgresql/shared_backups
     ports:
       - "5433:5432"
