+++ src/core/db/__init__.py
@@ -1,5 +1,9 @@
-from core.db.mixins import DateTimeMixin
+from core.db.mixins.time_mixin import TimeMixin
+from core.db.mixins.user_mixin import UserMixin
+from core.db.mixins.tree_mixin import TreeMixin
 from .models import Base
 
 
-__all__ = ["DateTimeMixin", "Base"]
+__all__ = ["TimeMixin",
+           "UserMixin",
+           "Base"]
