+++ src/core/db/mixins/time_mixin.py
@@ -1,7 +1,7 @@
 from sqlalchemy import orm, DateTime, func
 
 
-class DateTimeMixin:
+class TimeMixin:
     created_at = orm.mapped_column(DateTime(timezone=True), server_default=func.now())
     updated_at = orm.mapped_column(
         DateTime(timezone=True), server_default=func.now(), onupdate=func.now()
