--- src/core/factory/factory.py
@@ -1,334 +1,0 @@
-# from uuid import UUID
-#
-# from sqlalchemy import delete, select
-# from sqlalchemy.exc import IntegrityError
-# from sqlalchemy.ext.asyncio import AsyncSession
-#
-# from .models import SnippetModel
-# from categories.schemas import SnippetSchema
-#
-# class SnippetException(Exception):
-#     pass
-#
-#
-# class IntegrityConflictException(Exception):
-#     pass
-#
-#
-# class NotFoundException(Exception):
-#     pass
-#
-#
-# def CrudFactory(model: SnippetModel):
-#     class AsyncCrud:
-#         @classmethod
-#         async def create(
-#             cls,
-#             session: AsyncSession,
-#             data: SnippetSchema,
-#         ) -> SnippetModel:
-#             try:
-#                 db_model = model(**data.model_dump())
-#                 session.add(db_model)
-#                 await session.commit()
-#                 await session.refresh(db_model)
-#                 return db_model
-#             except IntegrityError:
-#                 raise IntegrityConflictException(
-#                     f"{model.__tablename__} conflicts with existing data.",
-#                 )
-#             except Exception as e:
-#                 raise SnippetException(f"Unknown error occurred: {e}") from e
-#
-#         @classmethod
-#         async def create_many(
-#             cls,
-#             session: AsyncSession,
-#             data: list[SnippetSchema],
-#             return_models: bool = False,
-#         ) -> list[SnippetModel] | bool:
-#             db_models = [model(**d.model_dump()) for d in data]
-#             try:
-#                 session.add_all(db_models)
-#                 await session.commit()
-#             except IntegrityError:
-#                 raise IntegrityConflictException(
-#                     f"{model.__tablename__} conflict with existing data.",
-#                 )
-#             except Exception as e:
-#                 raise SnippetException(f"Unknown error occurred: {e}") from e
-#
-#             if not return_models:
-#                 return True
-#
-#             for m in db_models:
-#                 await session.refresh(m)
-#
-#             return db_models
-#
-#         @classmethod
-#         async def get_one_by_id(
-#             cls,
-#             session: AsyncSession,
-#             id_: str | UUID,
-#             column: str = "uuid",
-#             with_for_update: bool = False,
-#         ) -> SnippetModel:
-#             try:
-#                 q = select(model).where(getattr(model, column) == id_)
-#             except AttributeError:
-#                 raise SnippetException(
-#                     f"Column {column} not found on {model.__tablename__}.",
-#                 )
-#
-#             if with_for_update:
-#                 q = q.with_for_update()
-#
-#             results = await session.execute(q)
-#             return results.unique().scalar_one_or_none()
-#
-#         @classmethod
-#         async def get_many_by_ids(
-#             cls,
-#             session: AsyncSession,
-#             ids: list[str | UUID] = None,
-#             column: str = "uuid",
-#             with_for_update: bool = False,
-#         ) -> list[SnippetModel]:
-#             q = select(model)
-#             if ids:
-#                 try:
-#                     q = q.where(getattr(model, column).in_(ids))
-#                 except AttributeError:
-#                     raise SnippetException(
-#                         f"Column {column} not found on {model.__tablename__}.",
-#                     )
-#
-#             if with_for_update:
-#                 q = q.with_for_update()
-#
-#             rows = await session.execute(q)
-#             return rows.unique().scalars().all()
-#
-#         @classmethod
-#         async def update_by_id(
-#             cls,
-#             session: AsyncSession,
-#             data: SnippetSchema,
-#             id_: str | UUID,
-#             column: str = "uuid",
-#         ) -> SnippetModel:
-#             db_model = await cls.get_one_by_id(
-#                 session, id_, column=column, with_for_update=True
-#             )
-#             if not db_model:
-#                 raise NotFoundException(
-#                     f"{model.__tablename__} {column}={id_} not found.",
-#                 )
-#
-#             values = data.model_dump(exclude_unset=True)
-#             for k, v in values.items():
-#                 setattr(db_model, k, v)
-#
-#             try:
-#                 await session.commit()
-#                 await session.refresh(db_model)
-#                 return db_model
-#             except IntegrityError:
-#                 raise IntegrityConflictException(
-#                     f"{model.__tablename__} {column}={id_} conflict with existing data.",
-#                 )
-#
-#         @classmethod
-#         async def update_many_by_ids(
-#             cls,
-#             session: AsyncSession,
-#             updates: dict[str | UUID, SnippetSchema],
-#             column: str = "uuid",
-#             return_models: bool = False,
-#         ) -> list[SnippetModel] | bool:
-#             updates = {str(id): update for id, update in updates.items() if update}
-#             ids = list(updates.keys())
-#             db_models = await cls.get_many_by_ids(
-#                 session, ids=ids, column=column, with_for_update=True
-#             )
-#
-#             for db_model in db_models:
-#                 values = updates[str(getattr(db_model, column))].model_dump(
-#                     exclude_unset=True
-#                 )
-#                 for k, v in values.items():
-#                     setattr(db_model, k, v)
-#                 session.add(db_model)
-#
-#             try:
-#                 await session.commit()
-#             except IntegrityError:
-#                 raise IntegrityConflictException(
-#                     f"{model.__tablename__} conflict with existing data.",
-#                 )
-#
-#             if not return_models:
-#                 return True
-#
-#             for db_model in db_models:
-#                 await session.refresh(db_model)
-#
-#             return db_models
-#
-#         @classmethod
-#         async def remove_by_id(
-#             cls,
-#             session: AsyncSession,
-#             id_: str | UUID,
-#             column: str = "uuid",
-#         ) -> int:
-#             try:
-#                 query = delete(model).where(getattr(model, column) == id_)
-#             except AttributeError:
-#                 raise SnippetException(
-#                     f"Column {column} not found on {model.__tablename__}.",
-#                 )
-#
-#             rows = await session.execute(query)
-#             await session.commit()
-#             return rows.rowcount
-#
-#         @classmethod
-#         async def remove_many_by_ids(
-#             cls,
-#             session: AsyncSession,
-#             ids: list[str | UUID],
-#             column: str = "uuid",
-#         ) -> int:
-#             if not ids:
-#                 raise SnippetException("No ids provided.")
-#
-#             try:
-#                 query = delete(model).where(getattr(model, column).in_(ids))
-#             except AttributeError:
-#                 raise SnippetException(
-#                     f"Column {column} not found on {model.__tablename__}.",
-#                 )
-#
-#             rows = await session.execute(query)
-#             await session.commit()
-#             return rows.rowcount
-#
-#     AsyncCrud.model = model
-from uuid import UUID
-
-from pydantic import BaseModel
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import DeclarativeMeta
-from sqlalchemy.exc import IntegrityError
-from sqlalchemy import select, delete
-
-# from repository import BaseRepository, ModelType
-
-
-# def CrudFactory(model: type[ModelType], session: AsyncSession) -> BaseRepository:
-#     """Фабрика, которая возвращает репозиторий для указанной модели."""
-#     return BaseRepository(model, session)
-
-
-ModelType = DeclarativeMeta
-SchemaType = BaseModel
-
-
-class SnippetException(Exception):
-    """Base exception for CRUD errors."""
-    def __init__(self, message: str):
-        super().__init__(message)
-
-class IntegrityConflictException(SnippetException):
-    """Raised when there is a database integrity conflict, such as duplicate entries."""
-    def __init__(self, message: str):
-        super().__init__(message)
-
-class NotFoundException(SnippetException):
-    """Raised when an entity is not found in the database."""
-    def __init__(self, message: str):
-        super().__init__(message)
-
-
-def CrudFactory(model: ModelType):
-    class AsyncCrud:
-        @classmethod
-        async def create(
-                cls,
-                session: AsyncSession,
-                data: SchemaType,
-        ) -> ModelType:
-            try:
-                obj = model(**data.model_dump())
-                session.add(obj)
-                await session.commit()
-                await session.refresh(obj)
-                return obj
-            except IntegrityError:
-                raise IntegrityConflictException(f"{cls.model.__tablename__} conflicts with existing data.")
-            # except Exception as e:
-            #     raise SnippetException(f"Unknown error occurred: {e}") from e
-
-        @classmethod
-        async def get_one_by_id(
-                cls,
-                session: AsyncSession,
-                id_: str | UUID,
-                column: str = "id",  # Supports both 'id' and 'uuid'
-                with_for_update: bool = False,
-        ) -> ModelType:
-            try:
-                q = select(model).where(getattr(model, column) == id_)
-            except AttributeError:
-                raise SnippetException(f"Column {column} not found on {model.__tablename__}.")
-
-            if with_for_update:
-                q = q.with_for_update()
-
-            results = await session.execute(q)
-            return results.unique().scalar_one_or_none()
-
-        @classmethod
-        async def update_by_id(
-                cls,
-                session: AsyncSession,
-                data: SchemaType,
-                id_: str | UUID,
-                column: str = "id",
-        ) -> ModelType:
-            db_model = await cls.get_one_by_id(session, id_, column=column, with_for_update=True)
-            if not db_model:
-                raise NotFoundException(f"{model.__tablename__} {column}={id_} not found.")
-
-            values = data.model_dump(exclude_unset=True)
-            for k, v in values.items():
-                setattr(db_model, k, v)
-
-            try:
-                await session.commit()
-                await session.refresh(db_model)
-                return db_model
-            except IntegrityError:
-                raise IntegrityConflictException(
-                    f"{model.__tablename__} {column}={id_} conflict with existing data.")
-
-        @classmethod
-        async def remove_by_id(
-                cls,
-                session: AsyncSession,
-                id_: str | UUID,
-                column: str = "id",
-        ) -> int:
-            try:
-                query = delete(model).where(getattr(model, column) == id_)
-            except AttributeError:
-                raise SnippetException(f"Column {column} not found on {model.__tablename__}.")
-
-            rows = await session.execute(query)
-            await session.commit()
-            return rows.rowcount
-
-    AsyncCrud.model = model
-    return AsyncCrud
