+++ src/products/__init__.py
@@ -1,0 +1,4 @@
+from .brands.models import Brand
+from .categories.models import Category
+from .countries.models import Country
+from .products.models import Product
