+++ src/products/categories/routers.py
@@ -1,10 +1,9 @@
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, status, Path
-from sqlalchemy.exc import IntegrityError
 
-from categories.repository import CategoryRepository, get_category_repository
-from categories.schemas import CategoryCreate, CategoryRead, CategoryPut, CategoryPatch
+from products.categories.repository import CategoryRepository
+from products.categories.schemas import CategoryCreateSchema, CategoryUpdateSchema, CategoryResponseSchema
 from core.exceptions import CategoryAlreadyCreated, \
 	CategoryOnThisLevelAlreadyCreated, CategoryNotFounded
 from core.exceptions.exp import CategoryUnprocessableEntity
@@ -15,15 +14,15 @@ router = APIRouter(
 )
 
 CategoryID = Annotated[int, Path(..., alias="id")]
-CategoryRepo = Annotated[CategoryRepository, Depends(get_category_repository)]
+CategoryRepo = Annotated[CategoryRepository, Depends(CategoryRepository.get_category_repository)]
 
 # Add validating errors
 # and read abt HTTPException
 # (abt convert ex's to HTTPException)
 
-@router.post("/categories/create", response_model=CategoryCreate, status_code=status.HTTP_201_CREATED)
+@router.post("/categories/create", response_model=CategoryCreateSchema, status_code=status.HTTP_201_CREATED)
 async def create_category(
-		data: CategoryCreate,
+		data: CategoryCreateSchema,
 		category_repo: CategoryRepo
 ):
 
@@ -35,22 +34,21 @@ async def create_category(
 		err_msg = str(e)
 
 		if '"categories_pkey"' in err_msg:
-			# print("DA EST ЗHE!!!")
 			raise CategoryAlreadyCreated(data.id, data.name)
 
 		if '"uq_name_parent_id"' in err_msg:
-			# print("DA EST ЗHE!!!")
 			raise CategoryOnThisLevelAlreadyCreated(data.parent_id, data.name)
 
 		if 'cannot have itself as its parent' in err_msg:
 			raise CategoryUnprocessableEntity(data.id, data.parent_id)
+
 		# Handle any unexpected errors and return a 400 status code
 		else:
 			raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
 			                    detail="An error occurred while getting the category")
 
 
-@router.get("/categories/get/{id}", response_model=CategoryRead, status_code=status.HTTP_200_OK)
+@router.get("/categories/get/{id}", response_model=CategoryResponseSchema, status_code=status.HTTP_200_OK)
 async def get_category_by_id(
 		category_id: CategoryID,
 		category_repo: CategoryRepo
@@ -61,29 +59,22 @@ async def get_category_by_id(
 		return category
 
 	except Exception as e:
-		# Handle 404 errors
 		err_msg = str(e)
-		print(e)
 
 		if 'Item not found' in err_msg:
 			raise CategoryNotFounded(category_id)
+
 		# Handle any unexpected errors and return a 400 status code
 		else:
 			print(f"Error {err_msg}")
 			raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
 			                    detail="An error occurred while getting the category")
 
-	# except Exception as e:
-		# Handle any unexpected errors and return a 400 status code
-	# 	print(f"Error {str(e)}")
-	# 	raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
-	# 	                    detail="An error occurred while getting the category")
 
-
-@router.put("/categories/replace/{id}", response_model=CategoryPut, status_code=status.HTTP_200_OK)
-async def replace_category(
+@router.put("/categories/update/{id}", response_model=CategoryUpdateSchema, status_code=status.HTTP_200_OK)
+async def update_category(
 		category_id: CategoryID,
-		data: CategoryPut,
+		data: CategoryUpdateSchema,
 		category_repo: CategoryRepo
 ):
 
@@ -91,58 +82,20 @@ async def replace_category(
 		category_updated = await category_repo.put(category_id, data)
 		return category_updated
 
-	# except ValueError as e:
-	# 	raise HTTPException(status_code=400, detail=str(e))
-
-	# except IntegrityError as e:
-	# 	err_msg = e.args[0]
-
-	# except ValidationError as e:
-	# 	raise HTTPException(status_code=e.status_code, detail=e.message)
-
 	except Exception as e:
-		# Handle 404 errors
 		err_msg = str(e)
-		# print(e)
 
 		if 'cannot have itself as its parent' in err_msg:
 			raise CategoryUnprocessableEntity(data.id, data.parent_id)
 
 		if 'Item not found' in err_msg:
 			raise CategoryNotFounded(category_id)
+
 		# Handle any unexpected errors and return a 400 status code
 		else:
 			print(f"Error {err_msg}")
 			raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
 			                    detail="An error occurred while replacing the category")
-
-
-@router.patch("/categories/update/{id}", response_model=CategoryPatch, status_code=status.HTTP_200_OK)
-async def update_category(
-		category_id: CategoryID,
-		data: CategoryPatch,
-		category_repo: CategoryRepo
-):
-
-	try:
-		category_updated = await category_repo.patch(category_id, data)
-		return category_updated
-
-	except Exception as e:
-		# Handle 404 errors
-		err_msg = str(e)
-		print(e)
-
-		if 'cannot have itself as its parent' in err_msg:
-			raise CategoryUnprocessableEntity(data.id, data.parent_id)
-
-		if 'Item not found' in err_msg:
-			raise CategoryNotFounded(category_id)
-		# Handle any unexpected errors and return a 400 status code
-		else:
-			print(f"Error {err_msg}")
-			raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
-			                    detail="An error occurred while updating the category")
 
 
 @router.delete("/categories/delete/{id}", status_code=status.HTTP_200_OK)
@@ -156,12 +109,11 @@ async def delete_category_by_id(
 		return {"message": "Category deleted successfully"}
 
 	except Exception as e:
-		# Handle 404 errors
 		err_msg = str(e)
-		print(e)
 
 		if 'Item not found' in err_msg:
 			raise CategoryNotFounded(category_id)
+
 		# Handle any unexpected errors and return a 400 status code
 		else:
 			print(f"Error {err_msg}")
